Remove duplication

+++ src/main/scala-2.12/org/apache/spark/sql/util/ArrowUtilsExposed.scala
@@ -32,7 +32,7 @@ object ArrowUtilsExposed {
       val intVector = new IntVector(smallInt.getName, rootAllocator)
       intVector.setValueCount(smallInt.getValueCount)
       (0 until smallInt.getValueCount).foreach {
-        case idx if (!smallInt.isNull(idx)) => intVector.set(idx, smallInt.get(idx))
+        case idx if (!smallInt.isNull(idx)) => intVector.set(idx, smallInt.get(idx).toInt)
         case idx                            => intVector.setNull(idx)
       }
